--- sample-anvil/src/main/java/com/airbnb/mvrx/sample/anvil/User.kt
@@ -1,2 +1,0 @@
-package com.airbnb.mvrx.sample.anvil
-
