+++ sample-anvil/src/main/java/com/airbnb/mvrx/sample/anvil/di/SingleIn.kt
@@ -1,6 +1,5 @@
 package com.airbnb.mvrx.sample.anvil.di
 
-
 import javax.inject.Scope
 import kotlin.reflect.KClass
