+++ sample-anvil/src/main/java/com/airbnb/mvrx/sample/anvil/feature/ExampleFeatureRepository.kt
@@ -1,0 +1,10 @@
+package com.airbnb.mvrx.sample.anvil.feature
+
+import com.airbnb.mvrx.sample.anvil.di.SingleIn
+import javax.inject.Inject
+
+@SingleIn(ExampleFeatureScope::class)
+class ExampleFeatureScopedRepository @Inject constructor() {
+    @Suppress("FunctionOnlyReturningConstant")
+    operator fun invoke() = "Example Feature"
+}
