+++ sample-anvil/src/main/res/layout/activity_main.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/rootContainer"
-    android:name="com.airbnb.mvrx.sample.anvil.ExampleFeatureFragment"
+    android:name="com.airbnb.mvrx.sample.anvil.feature.ExampleFeatureFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
